# src/backend/tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-powered tech news aggregator API. Fetches articles from RSS feeds and videos from YouTube, processes with AI summarization, and serves personalized feeds.",
    "title": "blips-ai-news",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ops/status": {
      "get": {
        "description": "Comprehensive operational status endpoint (requires ADMIN_API_KEY).\n\nReturns:\n- Request metrics (last 60 minutes): total requests, error rate, latencies\n- Connection pool stats: Redis and DB pool utilization\n- Service health indicators\n\nUse this for operational monitoring dashboards and alerting.",
        "operationId": "operational_status_ops_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operational Status"
      }
    }
  }
}